Cpplint

+++ template_cpp/src/include/udp.hpp
@@ -1,12 +1,13 @@
 #ifndef _UDP_H_
 #define _UDP_H_
 
-#include "common.hpp"
 #include <vector>
+
+#include "common.hpp"
 
 #define PAYLOAD_META_SIZE 8
 
-size_t get_node_idx_by_id(std::vector<node_t> &nodes, unsigned long id);
+size_t get_node_idx_by_id(std::vector<node_t> &nodes, uint32_t id);
 
 bool send_udp_payload(int sockfd, node_t *receiver, payload_t *payload,
                       ssize_t size);
